tweak git setup script

+++ 2_setup-git.sh
@@ -11,7 +11,7 @@ fi
 ##### 3. install git #####
 list=( zlib-devel gcc perl perl-ExtUtils-MakeMaker make build-essential tcl8.4 tk8.4 gettext )
 for item in ${list[@]}; do
-  GREPRESULT=sudo yum list installed | grep $item.x86_64
+  GREPRESULT=`sudo yum list installed | grep $item.x86_64`
   if [ "x$GREPRESULT" = "x" ]; then
     sudo yum -y install $item
   else
@@ -19,15 +19,13 @@ for item in ${list[@]}; do
   fi
 done
 
-# for tclsh failed; using unoptimized loading
-
 # in this section run as hash, no need sudo.(hamatta)
 mkdir -p /home/hash/work/src
 cd /home/hash/work/src
-wget http://git-core.googlecode.com/files/git-1.7.7.tar.gz
-tar xvzf git-1.7.7.tar.gz
+wget http://git-core.googlecode.com/files/git-1.7.10.1.tar.gz
+tar xvzf git-1.7.10.1.tar.gz
 
-cd git-1.7.7
+cd git-1.7.10.1
 
 # compile, and install.
 ./configure --prefix=/usr/local
@@ -45,9 +43,6 @@ fi
 git config --global user.name "takuyahashimoto"
 git config --global user.email takuya21hashimoto@gmail.com
 
-cd /home/hash
-mkdir git
-
 #########################################
 ########     GitHub settings    #########
 #########################################
@@ -61,11 +56,18 @@ if [ "`git --version | awk '{print $1,$2}'`" != "git version" ]; then
   exit 1
 fi
 
+cd /home/hash
 mkdir ~/.ssh
 
 echo "/////// enter keyname = id_rsa_github ///////"
 ssh-keygen -t rsa
-mv id_rsa_github* ~/.ssh/
+
+if [ -f id_rsa_github ];then
+  mv id_rsa_github* ~/.ssh/
+else
+  echo "create ssh key named 'id_rsa_github'. exit now."
+  exit 1
+fi
 
 cat <<EOF >> ~/.ssh/config
 ###################################
@@ -85,7 +87,7 @@ EOF
 chmod 700 ~/.ssh/
 chmod 600 ~/.ssh/config
 
-echo "/////// OK, now go to https://github.com/account ...
+echo "/////// OK, now go to https://github.com/settings/ssh
       /////// and insert text below."
 echo ""
 
@@ -99,12 +101,13 @@ read DUMMY
 echo "finish... now try to clone from GitHub"
 
 ### clone config
-cd /home/hash/git
-git clone --origin github git@github.com:takuyahashimoto/config.git
-echo "CMD :: git clone --origin github git@github.com:takuyahashimoto/config.git"
+cd /home/hash
+mkdir git && cd git
+git clone git@github.com:takuyahashimoto/config.git
+echo "CMD :: git clone git@github.com:takuyahashimoto/config.git"
 
 ### clone start-scripts
 cd /home/hash/git
-git clone --origin github git@github.com:takuyahashimoto/start-scripts.git
-echo "CMD :: git clone --origin github git@github.com:takuyahashimoto/start-scripts.git"
+git clone git@github.com:takuyahashimoto/start-scripts.git
+echo "CMD :: git clone git@github.com:takuyahashimoto/start-scripts.git"
 echo " now, you can run rest of start-scripts"
